[FIX] fzf-translate.sh: Limit context length in fzf-translate script
The  `fzf-translate.sh` script was modified to limit the amount of
context sent,
aiming to stay around 13.5k tokens.  A `max_chars` variable is set to
54000 and
`current_chars` keeps track of the current length.

The script now iterates through modified files and adds their diffs to
the
`context` variable. If adding a file's diff would exceed `max_chars`, a
warning
is appended to `context` indicating truncation and the loop breaks.