[FIX] .tmux.conf: Set default terminal to xterm-256color on Linux
The file `home/.tmux.conf` was modified to address an issue specific to Linux that caused `$TERM` to be set to `screen-256` instead of `tmux-256`, affecting the rendering of CommitCraft. A new line was added to set the default terminal to `xterm-256color` using the command `set -g default-terminal "xterm-256color"`. Additionally, an existing line was modified to include the `Tc` capability in the `terminal-overrides` option.