[IMP] install_commitCraft.sh: Update release tag and enforce overwrite in binary installation
The `tools/install_commitCraft.sh` script was modified to update the release tag from `v0.2.1` to `v0.2.3`. The `mv` commands used to move the binary to the installation directory were updated to use the `-f` flag, which forces the overwrite of the existing executable.

The script changes were limited to updating the release tag and adding the force overwrite functionality to the `mv` commands.