[FIX] .zshrc: Change the order of loading configurations
The file `home/.zshrc` was modified to change the order of loading configurations for Vim mode. The lines related to setting Vim mode (`set -o vi`) and loading the `zsh-vi-mode` plugin were moved to be above certain other configurations. Specifically, the `set -o vi` and the two `zinit` commands were relocated from the end to near the top of the file.