[IMP] atuin: Add new home directory keymap configurations
The file `home/.config/atuin/config.toml` was modified to add several new configurations. Specifically, the `keymap_mode` setting was uncommented and set to `"auto"`, and the `keymap_cursor` setting was uncommented and configured with cursor styles for different modes. Additionally, the `ignored_commands` setting was uncommented and configured to ignore the `"cd"` and `"ls"` commands.