[ADD] lazygit: commitcraft access in release mode
A new custom command was added to access commitcraft in release mode. The file `home/.config/lazygit/config.yml` was modified to include a new entry under `customCommands`. This command is triggered by the key `<c-c>` in the context of `commits`, and it executes the `commitcraft` command with the `-r` option via the subprocess `$HOME/.local/bin/commitcraft -r`. The command is described as 'Run commitcraft in Release mode'.