[FIX] .oh-my-posh: Fix cache configuration in shell prompt
The file `home/.oh-my-posh/prompt_config.toml` was modified to change the way the cache property is added. Specifically, the `cache_duration` property was replaced with a `cache` block that includes `duration` and `strategy` properties. The `duration` is set to `60s` and the `strategy` is set to `folder` in multiple segments. The change is applied to several blocks, including `properties.cache` in different segments.